cardinalBot.py
cardinalBot is a bot that automatically opens the game Cardinal in a webbrowser and then plays it. 
it uses mainly the pyautogui-module to locate images & pixels on the screen and control the game with up, down, left, right, keys. 
it was written on Winx64. 
Game Rules: Everytime one of the 4 surrounding blocks disappears (this is: the color changes from red to dark blue), the dot in the middle shall escape through the gap that is now opened.
Controls: up, down, left, right keys; space, click.

This bot was inspired after finishing Al Sweigat's Automate the boring Stuff as well as inventwithpython.com.

CH